Add Node.get_tree(), NodeTree.insert_before/after
Add node_tree_to dot() function.
When run as a script, generate 3 dot files, corresponding to the state
of a NodeTree before copies of sub-trees are inserted back into the
tree.  The names of the 3 generated files are:
    00_original.dot
    01_insert_1_after_0.dot
    02_insert_15_after_14.dot

The name of the file indicates the index of the root of the sub-tree
that was copied, along with the location where the sub-tree copy was
inserted.

Using dot (a program that is part of the GraphViz suite), the files can
be converted to PDF-rendered diagrams.

e.g.,
    for f in *.dot ; do dot -T pdf -o $(basename $f .dot).pdf $f; done;