Fix NodeTree grouping logic
Prior to this commit, the NodeTree grouping logic is invalid.  The basic
intent is to remove all but the first node from the list of selected
nodes to group, then append each of the removed nodes to the first node.
However, due to the previously mentioned logic error, the first node was
being removed instead, causing the entire list of nodes to be removed
from the tree, since the first node was the root for the group.